不具合対応とエーラメッセージ機能を追加

+++ src/main/java/com/marublosso/worktrack/worktrack_backend/dto/WorkTimeRequestDto.java
@@ -13,6 +13,6 @@ public class WorkTimeRequestDto {
     private LocalDateTime endTime;
     private Double totalHours;
     private Double overtime;
-    private int worktype;
+    private Long worktype;
 }
